Небольшие исправления в моделях

+++ EF_Console/AppContext.cs
@@ -5,15 +5,26 @@ namespace EF_Console
 {
     public class AppContext : DbContext
     {
+        private static bool createDB;
+
         public DbSet<User> Users { get; set; }
         public DbSet<Book> Books { get; set; }
         public DbSet<Author> Authors { get; set; }
         public DbSet<Genre> Genres { get; set; }
 
+        static AppContext()
+        {
+            createDB = true;
+        }
+
         public AppContext()
         {
-            Database.EnsureDeleted();
-            Database.EnsureCreated();
+            if(createDB)
+            {
+                createDB = false;
+                Database.EnsureDeleted();
+                Database.EnsureCreated();
+            }
         }
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
